% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{BucketOptions}
\alias{BucketOptions}
\title{BucketOptions Object}
\usage{
BucketOptions(exponentialBuckets = NULL, linearBuckets = NULL,
  explicitBuckets = NULL)
}
\arguments{
\item{exponentialBuckets}{The exponential buckets}

\item{linearBuckets}{The linear bucket}

\item{explicitBuckets}{The explicit buckets}
}
\value{
BucketOptions object
}
\description{
BucketOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Distribution may optionally contain a histogram of the values in the population. The histogram is given in bucket_counts as counts of values that fall into one of a sequence of non-overlapping buckets. The sequence of buckets is described by bucket_options.A bucket specifies an inclusive lower bound and exclusive upper bound for the values that are counted for that bucket. The upper bound of a bucket is strictly greater than the lower bound.The sequence of N buckets for a Distribution consists of an underflow bucket (number 0), zero or more finite buckets (number 1 through N - 2) and an overflow bucket (number N - 1). The buckets are contiguous: the lower bound of bucket i (i > 0) is the same as the upper bound of bucket i - 1. The buckets span the whole range of finite values: lower bound of the underflow bucket is -infinity and the upper bound of the overflow bucket is +infinity. The finite buckets are so-called because both bounds are finite.BucketOptions describes bucket boundaries in one of three ways. Two describe the boundaries by giving parameters for a formula to generate boundaries and one gives the bucket boundaries explicitly.If bucket_options is not given, then no bucket_counts may be given.
}
