% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{CollectdPayload.metadata}
\alias{CollectdPayload.metadata}
\title{CollectdPayload.metadata Object}
\usage{
CollectdPayload.metadata()
}
\value{
CollectdPayload.metadata object
}
\description{
CollectdPayload.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The measurement metadata. Example: `'process_id' -> 12345`
}
\seealso{
Other CollectdPayload functions: \code{\link{CollectdPayload}}
}

