% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{CollectdPayload}
\alias{CollectdPayload}
\title{CollectdPayload Object}
\usage{
CollectdPayload(CollectdPayload.metadata = NULL, values = NULL,
  startTime = NULL, endTime = NULL, plugin = NULL,
  pluginInstance = NULL, type = NULL, typeInstance = NULL,
  metadata = NULL)
}
\arguments{
\item{CollectdPayload.metadata}{The \link{CollectdPayload.metadata} object or list of objects}

\item{values}{The measured values during this time interval}

\item{startTime}{The start time of the interval}

\item{endTime}{The end time of the interval}

\item{plugin}{The name of the plugin}

\item{pluginInstance}{The instance name of the plugin Example: `'hdcl'`}

\item{type}{The measurement type}

\item{typeInstance}{The measurement type instance}

\item{metadata}{The measurement metadata}
}
\value{
CollectdPayload object
}
\description{
CollectdPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of data points sent from a `collectd`-based plugin. See the `collectd` documentation for more information.
}
\seealso{
Other CollectdPayload functions: \code{\link{CollectdPayload.metadata}}
}

