% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{CreateCollectdTimeSeriesRequest}
\alias{CreateCollectdTimeSeriesRequest}
\title{CreateCollectdTimeSeriesRequest Object}
\usage{
CreateCollectdTimeSeriesRequest(collectdVersion = NULL, resource = NULL,
  collectdPayloads = NULL)
}
\arguments{
\item{collectdVersion}{The version of collectd that collected the data}

\item{resource}{The monitored resource associated with the time series}

\item{collectdPayloads}{The collectd payloads representing the time series data}
}
\value{
CreateCollectdTimeSeriesRequest object
}
\description{
CreateCollectdTimeSeriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The CreateCollectdTimeSeries request.
}
\seealso{
Other CreateCollectdTimeSeriesRequest functions: \code{\link{projects.collectdTimeSeries.create}}
}
