% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{CreateTimeSeriesRequest}
\alias{CreateTimeSeriesRequest}
\title{CreateTimeSeriesRequest Object}
\usage{
CreateTimeSeriesRequest(timeSeries = NULL)
}
\arguments{
\item{timeSeries}{The new data to be added to a list of time series}
}
\value{
CreateTimeSeriesRequest object
}
\description{
CreateTimeSeriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The `CreateTimeSeries` request.
}
\seealso{
Other CreateTimeSeriesRequest functions: \code{\link{projects.timeSeries.create}}
}

