% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Distribution}
\alias{Distribution}
\title{Distribution Object}
\usage{
Distribution(count = NULL, mean = NULL, sumOfSquaredDeviation = NULL,
  range = NULL, bucketOptions = NULL, bucketCounts = NULL)
}
\arguments{
\item{count}{The number of values in the population}

\item{mean}{The arithmetic mean of the values in the population}

\item{sumOfSquaredDeviation}{The sum of squared deviations from the mean of the values in the population}

\item{range}{If specified, contains the range of the population values}

\item{bucketOptions}{Defines the histogram bucket boundaries}

\item{bucketCounts}{If `bucket_options` is given, then the sum of the values in `bucket_counts` must equal the value in `count`}
}
\value{
Distribution object
}
\description{
Distribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Distribution contains summary statistics for a population of values and, optionally, a histogram representing the distribution of those values across a specified set of histogram buckets. The summary statistics are the count, mean, sum of the squared deviation from the mean, the minimum, and the maximum of the set of population of values. The histogram is based on a sequence of buckets and gives a count of values that fall into each bucket. The boundaries of the buckets are given either explicitly or by specifying parameters for a method of computing them (buckets of fixed width or buckets of exponentially increasing width). Although it is not forbidden, it is generally a bad idea to include non-finite values (infinities or NaNs) in the population of values, as this will render the `mean` and `sum_of_squared_deviation` fields meaningless.
}

