% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Empty}
\alias{Empty}
\title{Empty Object}
\usage{
Empty()
}
\value{
Empty object
}
\description{
Empty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:service Foo {  rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);}The JSON representation for Empty is empty JSON object {}.
}
