% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Explicit}
\alias{Explicit}
\title{Explicit Object}
\usage{
Explicit(bounds = NULL)
}
\arguments{
\item{bounds}{The values must be monotonically increasing}
}
\value{
Explicit object
}
\description{
Explicit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of buckets with arbitrary widths.Defines size(bounds) + 1 (= N) buckets with these boundaries for bucket i:Upper bound (0 <= i < N-1): boundsi  Lower bound (1 <= i < N); boundsi - 1There must be at least one element in bounds. If bounds has only one element, there are no finite buckets, and that single element is the common boundary of the overflow and underflow buckets.
}
