% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Object}
\usage{
Exponential(numFiniteBuckets = NULL, growthFactor = NULL, scale = NULL)
}
\arguments{
\item{numFiniteBuckets}{must be greater than 0}

\item{growthFactor}{Must be greater than 1}

\item{scale}{Must be greater than 0}
}
\value{
Exponential object
}
\description{
Exponential Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specify a sequence of buckets that have a width that is proportional to the value of the lower bound. Each bucket represents a constant relative uncertainty on a specific value in the bucket. Defines `num_finite_buckets + 2` (= N) buckets with these boundaries for bucket i: Upper bound (0 <= i < N-1): scale * (growth_factor ^ i). Lower bound (1 <= i < N): scale * (growth_factor ^ (i - 1)).
}

