% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Linear}
\alias{Linear}
\title{Linear Object}
\usage{
Linear(numFiniteBuckets = NULL, width = NULL, offset = NULL)
}
\arguments{
\item{numFiniteBuckets}{Must be greater than 0}

\item{width}{Must be greater than 0}

\item{offset}{Lower bound of the first bucket}
}
\value{
Linear object
}
\description{
Linear Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specify a sequence of buckets that all have the same width (except overflow and underflow). Each bucket represents a constant absolute uncertainty on the specific value in the bucket. Defines `num_finite_buckets + 2` (= N) buckets with these boundaries for bucket `i`: Upper bound (0 <= i < N-1): offset + (width * i). Lower bound (1 <= i < N): offset + (width * (i - 1)).
}

