% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{ListGroupMembersResponse}
\alias{ListGroupMembersResponse}
\title{ListGroupMembersResponse Object}
\usage{
ListGroupMembersResponse(members = NULL, nextPageToken = NULL,
  totalSize = NULL)
}
\arguments{
\item{members}{A set of monitored resources in the group}

\item{nextPageToken}{If there are more results than have been returned, then this field is set to a non-empty value}

\item{totalSize}{The total number of elements matching this request}
}
\value{
ListGroupMembersResponse object
}
\description{
ListGroupMembersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The `ListGroupMembers` response.
}

