% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{ListGroupsResponse}
\alias{ListGroupsResponse}
\title{ListGroupsResponse Object}
\usage{
ListGroupsResponse(group = NULL, nextPageToken = NULL)
}
\arguments{
\item{group}{The groups that match the specified filters}

\item{nextPageToken}{If there are more results than have been returned, then this field is set to a non-empty value}
}
\value{
ListGroupsResponse object
}
\description{
ListGroupsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The ListGroups response.
}
