% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{ListTimeSeriesResponse}
\alias{ListTimeSeriesResponse}
\title{ListTimeSeriesResponse Object}
\usage{
ListTimeSeriesResponse(timeSeries = NULL, nextPageToken = NULL)
}
\arguments{
\item{timeSeries}{One or more time series that match the filter included in the request}

\item{nextPageToken}{If there are more results than have been returned, then this field is set to a non-empty value}
}
\value{
ListTimeSeriesResponse object
}
\description{
ListTimeSeriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The ListTimeSeries response.
}
