% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Metric}
\alias{Metric}
\title{Metric Object}
\usage{
Metric(Metric.labels = NULL, type = NULL, labels = NULL)
}
\arguments{
\item{Metric.labels}{The \link{Metric.labels} object or list of objects}

\item{type}{An existing metric type, see google}

\item{labels}{The set of labels that uniquely identify a metric}
}
\value{
Metric object
}
\description{
Metric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A specific metric identified by specifying values for all of the labels of a `MetricDescriptor`.
}
\seealso{
Other Metric functions: \code{\link{Metric.labels}}
}

