% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{MetricDescriptor}
\alias{MetricDescriptor}
\title{MetricDescriptor Object}
\usage{
MetricDescriptor(name = NULL, type = NULL, labels = NULL,
  metricKind = NULL, valueType = NULL, unit = NULL, description = NULL,
  displayName = NULL)
}
\arguments{
\item{name}{Resource name}

\item{type}{The metric type including a DNS name prefix, for example `'compute}

\item{labels}{The set of labels that can be used to describe a specific instance of this metric type}

\item{metricKind}{Whether the metric records instantaneous values, changes to a value, etc}

\item{valueType}{Whether the measurement is an integer, a floating-point number, etc}

\item{unit}{The unit in which the metric value is reported}

\item{description}{A detailed description of the metric, which can be used in documentation}

\item{displayName}{A concise name for the metric, which can be displayed in user interfaces}
}
\value{
MetricDescriptor object
}
\description{
MetricDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a metric type and its schema.
}
\seealso{
Other MetricDescriptor functions: \code{\link{projects.metricDescriptors.create}}
}

