% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_functions.R
\docType{package}
\name{monitoring_googleAuthR}
\alias{monitoring_googleAuthR}
\alias{monitoring_googleAuthR-package}
\title{Stackdriver Monitoring API
Manages your Stackdriver Monitoring data and configurations. Most projects must be associated with a Stackdriver account, with a few exceptions as noted on the individual method pages.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:02:52
filename: /Users/mark/dev/R/autoGoogleAPI/googlemonitoringv3.auto/R/monitoring_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring.read
\item https://www.googleapis.com/auth/monitoring
\item https://www.googleapis.com/auth/monitoring.write
}
}
