% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Option.value}
\alias{Option.value}
\title{Option.value Object}
\usage{
Option.value()
}
\value{
Option.value object
}
\description{
Option.value Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The option's value. For example, `'com.google.protobuf'`.
}
\seealso{
Other Option functions: \code{\link{Option}}
}

