% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Point}
\alias{Point}
\title{Point Object}
\usage{
Point(interval = NULL, value = NULL)
}
\arguments{
\item{interval}{The time interval to which the data point applies}

\item{value}{The value of the data point}
}
\value{
Point object
}
\description{
Point Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single data point in a time series.
}
