% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_functions.R
\name{projects.collectdTimeSeries.create}
\alias{projects.collectdTimeSeries.create}
\title{Creates a new time series with the given data points. This method is only for use in `collectd`-related code, including the Google Monitoring Agent. See [google.monitoring.v3.MetricService.CreateTimeSeries] instead.}
\usage{
projects.collectdTimeSeries.create(CreateCollectdTimeSeriesRequest, name)
}
\arguments{
\item{CreateCollectdTimeSeriesRequest}{The \link{CreateCollectdTimeSeriesRequest} object to pass to this method}

\item{name}{The project in which to create the time series}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
\item https://www.googleapis.com/auth/monitoring.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/monitoring, https://www.googleapis.com/auth/monitoring.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/monitoring/api/}{Google Documentation}

Other CreateCollectdTimeSeriesRequest functions: \code{\link{CreateCollectdTimeSeriesRequest}}
}

