% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_functions.R
\name{projects.metricDescriptors.get}
\alias{projects.metricDescriptors.get}
\title{Gets a single metric descriptor. This method does not require a Stackdriver account.}
\usage{
projects.metricDescriptors.get(name)
}
\arguments{
\item{name}{The metric descriptor on which to execute the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
\item https://www.googleapis.com/auth/monitoring.read
\item https://www.googleapis.com/auth/monitoring.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/monitoring, https://www.googleapis.com/auth/monitoring.read, https://www.googleapis.com/auth/monitoring.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/monitoring/api/}{Google Documentation}
}

