% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_functions.R
\name{projects.timeSeries.create}
\alias{projects.timeSeries.create}
\title{Creates or adds data to one or more time series. The response is empty if all time series in the request were written. If any time series could not be written, a corresponding failure message is included in the error response.}
\usage{
projects.timeSeries.create(CreateTimeSeriesRequest, name)
}
\arguments{
\item{CreateTimeSeriesRequest}{The \link{CreateTimeSeriesRequest} object to pass to this method}

\item{name}{The project on which to execute the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
\item https://www.googleapis.com/auth/monitoring.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/monitoring, https://www.googleapis.com/auth/monitoring.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/monitoring/api/}{Google Documentation}

Other CreateTimeSeriesRequest functions: \code{\link{CreateTimeSeriesRequest}}
}
