% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_functions.R
\name{projects.timeSeries.list}
\alias{projects.timeSeries.list}
\title{Lists time series that match a filter. This method does not require a Stackdriver account.}
\usage{
projects.timeSeries.list(name, orderBy = NULL,
  aggregation.crossSeriesReducer = NULL, filter = NULL,
  aggregation.perSeriesAligner = NULL, pageToken = NULL,
  interval.startTime = NULL, view = NULL,
  aggregation.groupByFields = NULL, interval.endTime = NULL,
  aggregation.alignmentPeriod = NULL, pageSize = NULL)
}
\arguments{
\item{name}{The project on which to execute the request}

\item{orderBy}{Specifies the order in which the points of the time series should be returned}

\item{aggregation.crossSeriesReducer}{The approach to be used to combine time series}

\item{filter}{A monitoring filter that specifies which time series should be returned}

\item{aggregation.perSeriesAligner}{The approach to be used to align individual time series}

\item{pageToken}{If this field is not empty then it must contain the nextPageToken value returned by a previous call to this method}

\item{interval.startTime}{Optional}

\item{view}{Specifies which information is returned about the time series}

\item{aggregation.groupByFields}{The set of fields to preserve when crossSeriesReducer is specified}

\item{interval.endTime}{Required}

\item{aggregation.alignmentPeriod}{The alignment period for per-time series alignment}

\item{pageSize}{A positive number that is the maximum number of results to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
\item https://www.googleapis.com/auth/monitoring.read
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/monitoring, https://www.googleapis.com/auth/monitoring.read)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/monitoring/api/}{Google Documentation}
}
