% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Range}
\alias{Range}
\title{Range Object}
\usage{
Range(max = NULL, min = NULL)
}
\arguments{
\item{max}{The maximum of the population values}

\item{min}{The minimum of the population values}
}
\value{
Range object
}
\description{
Range Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The range of the population values.
}
