% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{SourceContext}
\alias{SourceContext}
\title{SourceContext Object}
\usage{
SourceContext(fileName = NULL)
}
\arguments{
\item{fileName}{The path-qualified name of the}
}
\value{
SourceContext object
}
\description{
SourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SourceContext represents information about the source of a protobuf element, like the file in which it is defined.
}
