% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{TimeInterval}
\alias{TimeInterval}
\title{TimeInterval Object}
\usage{
TimeInterval(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{Required}

\item{startTime}{Optional}
}
\value{
TimeInterval object
}
\description{
TimeInterval Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A time interval extending just after a start time through an end time. If the start time is the same as the end time, then the interval represents a single point in time.
}
