% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{TimeSeries}
\alias{TimeSeries}
\title{TimeSeries Object}
\usage{
TimeSeries(resource = NULL, metricKind = NULL, metric = NULL,
  points = NULL, valueType = NULL)
}
\arguments{
\item{resource}{The associated resource}

\item{metricKind}{The metric kind of the time series}

\item{metric}{The associated metric}

\item{points}{The data points of this time series}

\item{valueType}{The value type of the time series}
}
\value{
TimeSeries object
}
\description{
TimeSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of data points that describes the time-varying values of a metric. A time series is identified by a combination of a fully-specified monitored resource and a fully-specified metric. This type is used for both listing and creating time series.
}
