% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{TypedValue}
\alias{TypedValue}
\title{TypedValue Object}
\usage{
TypedValue(boolValue = NULL, int64Value = NULL, doubleValue = NULL,
  stringValue = NULL, distributionValue = NULL)
}
\arguments{
\item{boolValue}{A Boolean value: `true` or `false`}

\item{int64Value}{A 64-bit integer}

\item{doubleValue}{A 64-bit double-precision floating-point number}

\item{stringValue}{A variable-length string value}

\item{distributionValue}{A distribution value}
}
\value{
TypedValue object
}
\description{
TypedValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single strongly-typed value.
}

