% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{CollectdValue}
\alias{CollectdValue}
\title{CollectdValue Object}
\usage{
CollectdValue(value = NULL, dataSourceType = NULL, dataSourceName = NULL)
}
\arguments{
\item{value}{The measurement value}

\item{dataSourceType}{The type of measurement}

\item{dataSourceName}{The data source for the collectd value}
}
\value{
CollectdValue object
}
\description{
CollectdValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single data point from a collectd-based plugin.
}
