% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{CreateCollectdTimeSeriesRequest}
\alias{CreateCollectdTimeSeriesRequest}
\title{Google Monitoring API Objects 
Manages your Stackdriver monitoring data and configurations. Projects must be associated with a Stackdriver account, except for the following methods: [monitoredResourceDescriptors.list](v3/projects.monitoredResourceDescriptors/list), [monitoredResourceDescriptors.get](v3/projects.monitoredResourceDescriptors/get), [metricDescriptors.list](v3/projects.metricDescriptors/list), [metricDescriptors.get](v3/projects.metricDescriptors/get), and [timeSeries.list](v3/projects.timeSeries/list).}
\usage{
CreateCollectdTimeSeriesRequest(resource = NULL, collectdVersion = NULL,
  collectdPayloads = NULL)
}
\arguments{
\item{resource}{The monitored resource associated with the time series}

\item{collectdVersion}{The version of `collectd` that collected the data}

\item{collectdPayloads}{The `collectd` payloads representing the time series data}
}
\value{
CreateCollectdTimeSeriesRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:20:45
filename: /Users/mark/dev/R/autoGoogleAPI/googlemonitoringv3.auto/R/monitoring_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
CreateCollectdTimeSeriesRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The `CreateCollectdTimeSeries` request.
}
\seealso{
Other CreateCollectdTimeSeriesRequest functions: \code{\link{projects.collectdTimeSeries.create}}
}

