% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{ListGroupMembersResponse}
\alias{ListGroupMembersResponse}
\title{Stackdriver Monitoring API Objects 
Manages your Stackdriver Monitoring data and configurations. Most projects must be associated with a Stackdriver account, with a few exceptions as noted on the individual method pages.}
\usage{
ListGroupMembersResponse(totalSize = NULL, members = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{totalSize}{The total number of elements matching this request}

\item{members}{A set of monitored resources in the group}

\item{nextPageToken}{If there are more results than have been returned, then this field is set to a non-empty value}
}
\value{
ListGroupMembersResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:02:52
filename: /Users/mark/dev/R/autoGoogleAPI/googlemonitoringv3.auto/R/monitoring_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListGroupMembersResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The ListGroupMembers response.
}
