% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{ListMonitoredResourceDescriptorsResponse}
\alias{ListMonitoredResourceDescriptorsResponse}
\title{ListMonitoredResourceDescriptorsResponse Object}
\usage{
ListMonitoredResourceDescriptorsResponse(resourceDescriptors = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{resourceDescriptors}{The monitored resource descriptors that are available to this project and that match `filter`, if present}

\item{nextPageToken}{If there are more results than have been returned, then this field is set to a non-empty value}
}
\value{
ListMonitoredResourceDescriptorsResponse object
}
\description{
ListMonitoredResourceDescriptorsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The `ListMonitoredResourcDescriptors` response.
}

