% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{Metric.labels}
\alias{Metric.labels}
\title{Metric.labels Object}
\usage{
Metric.labels()
}
\value{
Metric.labels object
}
\description{
Metric.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The set of labels that uniquely identify a metric. To specify a metric, all labels enumerated in the `MetricDescriptor` must be assigned values.
}
\seealso{
Other Metric functions: \code{\link{Metric}}
}

