% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{MetricDescriptor}
\alias{MetricDescriptor}
\title{MetricDescriptor Object}
\usage{
MetricDescriptor(labels = NULL, name = NULL, type = NULL,
  valueType = NULL, metricKind = NULL, description = NULL,
  displayName = NULL, unit = NULL)
}
\arguments{
\item{labels}{The set of labels that can be used to describe a specific instance of this metric type}

\item{name}{The resource name of the metric descriptor}

\item{type}{The metric type, including its DNS name prefix}

\item{valueType}{Whether the measurement is an integer, a floating-point number, etc}

\item{metricKind}{Whether the metric records instantaneous values, changes to a value, etc}

\item{description}{A detailed description of the metric, which can be used in documentation}

\item{displayName}{A concise name for the metric, which can be displayed in user interfaces}

\item{unit}{The unit in which the metric value is reported}
}
\value{
MetricDescriptor object
}
\description{
MetricDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a metric type and its schema. Once a metric descriptor is created, deleting or altering it stops data collection and makes the metric type's existing data unusable.
}
\seealso{
Other MetricDescriptor functions: \code{\link{projects.metricDescriptors.create}}
}
