% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_functions.R
\docType{package}
\name{monitoring_googleAuthR}
\alias{monitoring_googleAuthR}
\alias{monitoring_googleAuthR-package}
\title{Google Monitoring API
Manages your Stackdriver monitoring data and configurations. Projects must be associated with a Stackdriver account, except for the following methods: [monitoredResourceDescriptors.list](v3/projects.monitoredResourceDescriptors/list), [monitoredResourceDescriptors.get](v3/projects.monitoredResourceDescriptors/get), [metricDescriptors.list](v3/projects.metricDescriptors/list), [metricDescriptors.get](v3/projects.metricDescriptors/get), and [timeSeries.list](v3/projects.timeSeries/list).}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:20:44
filename: /Users/mark/dev/R/autoGoogleAPI/googlemonitoringv3.auto/R/monitoring_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
\item https://www.googleapis.com/auth/monitoring.read
\item https://www.googleapis.com/auth/monitoring.write
}
}

