% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{TimeInterval}
\alias{TimeInterval}
\title{TimeInterval Object}
\usage{
TimeInterval(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{Required}

\item{startTime}{If this value is omitted, the interval is a point in time, `endTime`}
}
\value{
TimeInterval object
}
\description{
TimeInterval Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A time interval extending from after `startTime` through `endTime`. If `startTime` is omitted, the interval is the single point in time, `endTime`.
}

