% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{TimeSeries}
\alias{TimeSeries}
\title{TimeSeries Object}
\usage{
TimeSeries(metric = NULL, resource = NULL, metricKind = NULL,
  valueType = NULL, points = NULL)
}
\arguments{
\item{metric}{The fully-specified metric used to identify the time series}

\item{resource}{The fully-specified monitored resource used to identify the time series}

\item{metricKind}{The metric kind of the time series}

\item{valueType}{The value type of the time series}

\item{points}{The data points of this time series}
}
\value{
TimeSeries object
}
\description{
TimeSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of data points that describes the time-varying nature of a metric. A time series is identified by a combination of a fully-specified monitored resource and a fully-specified metric.
}

