% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_objects.R
\name{TypedValue}
\alias{TypedValue}
\title{TypedValue Object}
\usage{
TypedValue(int64Value = NULL, distributionValue = NULL, boolValue = NULL,
  stringValue = NULL, doubleValue = NULL)
}
\arguments{
\item{int64Value}{A 64-bit integer}

\item{distributionValue}{A distribution value}

\item{boolValue}{A Boolean value: true or false}

\item{stringValue}{A variable-length string value}

\item{doubleValue}{A 64-bit double-precision floating-point number}
}
\value{
TypedValue object
}
\description{
TypedValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single strongly-typed value.
}
